requirements.txt
amqp==2.6.1
argh==0.26.2
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
black==22.1.0
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.12
django-appconf==1.0.5
django-autoslug==1.9.8
django-celery-email==3.0.0
django-countries==7.2.1
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==6.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
factory-boy==3.2.1
Faker==13.3.0
flake8==4.0.1
flower==0.9.7
humanize==4.0.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
Markdown==3.3.6
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.43
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.8.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
vine==1.3.0
watchdog==1.0.2
zipp==3.7.0
